Add new SVG icon to AzureCosmosDB.svg (#18)
Added a new SVG element with the ID `b089cfca-0de1-451c-a1ca-6680ea50cb4f` to the `AzureCosmosDB.svg` file. This SVG includes a radial gradient, a clip path, and multiple paths with different fill colors to create a complex icon design. The SVG is defined with a width and height of 18 units and a viewBox of `0 0 18 18`.